Sync LeetCode submission - Sum of Two Integers (cpp)

+++ problems/sum_of_two_integers/solution.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
-    int getSum(int a, int b) {
+    int getSum(unsigned int a, unsigned int b) {
         if(!b) return a;
-        return getSum(a ^ b, (unsigned int)(a & b) << 1);
+        return getSum(a ^ b, (a & b) << 1);
     }
 };
